+++ base/exports.jl
@@ -449,13 +449,14 @@ export
     ≉,
 
 # specfun
-    airy,
     airyai,
     airyaiprime,
     airybi,
     airybiprime,
-    airyprime,
-    airyx,
+    airyaix,
+    airyaiprimex,
+    airybix,
+    airybiprimex,
     besselh,
     besselhx,
     besseli,